Add files via upload
A program which can draw circle, rectangle, triangle or a dotted line if you press 1, 2, 3 or 4 on your keyboard.
Shapes can be resized by dragging with the mouse. 
Whatever you draw is stored in a text file and id displayed the next time the program is run.
Shapes are stored in stacks and put into a queue when removed by pressing the right click.
The are redrawn if you press the middle mouse button, one at a time.